Major relocation to the com.e_gineering package structure
