+++ src/main/java/com/e_gineering/Connection.java
@@ -1,4 +1,4 @@
-package org.collectd;
+package com.e_gineering;
 
 import javax.management.InstanceNotFoundException;
 import javax.management.ListenerNotFoundException;
